Remove the ci/cd distinction
lines between CI/CD are a little blurred, so I think it makes sense to have a
single directory. I've named it `pipeline` to indicate these scripts are
intended to be used when defining pipelines. They do not necessarily contain
full pipelines, but can be used as fragments in a larger pipeline.