Improve fact generation for config-diff
The issue was that when kolla-ansible and kayobe switched to reference facts in the
ansible_facts dictionary instead of the via the implicitly injected facts, our
fact overrides did not take effect. The new method is to generate json files for
the fact cache which should support implicit facts as well.

Conflicts:
ansible/roles/config-diff-vars/templates/kolla-extra-vars.yml.j2
docker/kayobe/Dockerfile