Allow skipping copy of kayobe-config
If kayobe-config is mounted at
/stack/kayobe-automation-env/src/kayobe-config, we can avoid copying the
configuration into place and instead work in the bind mounted
repository. This has the effect that changes made to the configuration
will be reflected in the original repository on the host. This may be
useful if we want to extract changes without configuring
kayobe-automation to create a PR.