Workaround broken rsync package (#24)
Workaround broken rsync package

The issue we are seeing is:

```
[root@ac99140be16b /]# /usr/bin/rsync -F --compress --recursive --rsh=/usr/bin/ssh stack@192.168.45.182:/tmp/a /stack/tempest-artifacts
rsync: on remote machine: --sparse-block=1024: unknown option
```

It has been reported here: https://bugzilla.redhat.com/show_bug.cgi?id=2043753 and a fix is being worked on. Obviously this change has a limited lifetime, but the alternative of using `--sparse-block=0`, has issues too. For example, it won't work on rsync versions that lack this patch. The plan is to revert this change once a new version of rsync, that isn't affected by this bug, has been released.