fix: handle encrypted `.yml` files correctly
Before `.yml` files were being grouped together with all other vaulted files. This is wrong as these `.yml` files need to be used by Ansible as files that contain variables.

Therefore, this change aims to split all encrypted files based on `.yml` extension or not. Also it has been identified that with this approach of find all vaulted files we are unintentionally pickup files from `ci-multinode` and other environments which can not be decrypted using client vault password.