fix: use correct path when normalising files
Due to a previous addition of adding support for parallel configuration generation when producing a config-diff there is an unfortunate side effect whereby some files such as fluentd contain the different paths they were produced from. This causes false positives in config-diff.

The change had included a fix however the wrong folder to search was specified.